+++ app/models/chip.py
@@ -1,7 +1,7 @@
 import enum
 from datetime import datetime
 
-from sqlalchemy import BigInteger, DateTime, Enum, ForeignKey, Integer, String, func
+from sqlalchemy import DateTime, Enum, ForeignKey, Integer, String, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.database import Base
@@ -20,14 +20,14 @@ class LedgerReasonType(str, enum.Enum):
 class ChipLedger(Base):
     __tablename__ = "chip_ledger"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
-    account_id: Mapped[int] = mapped_column(BigInteger, ForeignKey("accounts.id"), nullable=False)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
+    account_id: Mapped[int] = mapped_column(Integer, ForeignKey("accounts.id"), nullable=False)
     delta: Mapped[int] = mapped_column(Integer, nullable=False)
     balance_after: Mapped[int] = mapped_column(Integer, nullable=False)
     reason_type: Mapped[LedgerReasonType] = mapped_column(Enum(LedgerReasonType), nullable=False)
     reason_text: Mapped[str | None] = mapped_column(String(256), nullable=True)
     ref_type: Mapped[str | None] = mapped_column(String(64), nullable=True)
-    ref_id: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
+    ref_id: Mapped[int | None] = mapped_column(Integer, nullable=True)
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), server_default=func.now(), nullable=False)
 
     account: Mapped["Account"] = relationship(back_populates="ledger_entries")
