+++ app/models/credential.py
@@ -1,7 +1,7 @@
 import enum
 from datetime import datetime
 
-from sqlalchemy import BigInteger, DateTime, Enum, ForeignKey, String, UniqueConstraint, func
+from sqlalchemy import DateTime, Enum, ForeignKey, Integer, String, UniqueConstraint, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.database import Base
@@ -15,8 +15,8 @@ class CredentialStatus(str, enum.Enum):
 class ApiCredential(Base):
     __tablename__ = "api_credentials"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
-    account_id: Mapped[int] = mapped_column(BigInteger, ForeignKey("accounts.id"), nullable=False)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
+    account_id: Mapped[int] = mapped_column(Integer, ForeignKey("accounts.id"), nullable=False)
     api_key: Mapped[str] = mapped_column(String(128), unique=True, nullable=False)
     secret_hash: Mapped[str] = mapped_column(String(256), nullable=False)
     status: Mapped[CredentialStatus] = mapped_column(
@@ -33,7 +33,7 @@ class ApiNonce(Base):
     __tablename__ = "api_nonces"
     __table_args__ = (UniqueConstraint("api_key", "nonce", name="uq_api_nonces_key_nonce"),)
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     api_key: Mapped[str] = mapped_column(String(128), nullable=False, index=True)
     nonce: Mapped[str] = mapped_column(String(64), nullable=False)
-    timestamp: Mapped[int] = mapped_column(BigInteger, nullable=False)
+    timestamp: Mapped[int] = mapped_column(Integer, nullable=False)
