+++ app/models/table.py
@@ -1,7 +1,7 @@
 import enum
 from datetime import datetime
 
-from sqlalchemy import BigInteger, DateTime, Enum, ForeignKey, Integer, String, UniqueConstraint, func
+from sqlalchemy import DateTime, Enum, ForeignKey, Integer, String, UniqueConstraint, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.database import Base
@@ -22,7 +22,7 @@ class SeatStatus(str, enum.Enum):
 class Table(Base):
     __tablename__ = "tables"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     table_no: Mapped[int] = mapped_column(Integer, unique=True, nullable=False)
     status: Mapped[TableStatus] = mapped_column(Enum(TableStatus), default=TableStatus.OPEN, nullable=False)
     max_seats: Mapped[int] = mapped_column(Integer, default=9, nullable=False)
@@ -43,10 +43,10 @@ class TableSeat(Base):
     __tablename__ = "table_seats"
     __table_args__ = (UniqueConstraint("table_id", "seat_no", name="uq_table_seat"),)
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
-    table_id: Mapped[int] = mapped_column(BigInteger, ForeignKey("tables.id"), nullable=False)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
+    table_id: Mapped[int] = mapped_column(Integer, ForeignKey("tables.id"), nullable=False)
     seat_no: Mapped[int] = mapped_column(Integer, nullable=False)
-    account_id: Mapped[int | None] = mapped_column(BigInteger, ForeignKey("accounts.id"), nullable=True)
+    account_id: Mapped[int | None] = mapped_column(Integer, ForeignKey("accounts.id"), nullable=True)
     seat_status: Mapped[SeatStatus] = mapped_column(Enum(SeatStatus), default=SeatStatus.EMPTY, nullable=False)
     stack: Mapped[int] = mapped_column(Integer, default=0, nullable=False)
     joined_at: Mapped[datetime | None] = mapped_column(DateTime(timezone=True), nullable=True)
